[wip] Initial implementation of duplex for jack
It almost compiles.

Original (outdated) commit message:

The way the temporary interlacing buffers are handled currently is not
great. The root of the issue seems to be that JACK actually provides a
callback that allows performing allocations on the realtime thread when
buffer size changes, but this is not exposed as such in rust-jack.

See: https://github.com/RustAudio/rust-jack/issues/137

The duplex implementation now mirrors input, and this panicks if buffer
size becomes greater than the original.